move safeStorageService to electron subService directory
